Add --stdout option to print downloaded file to stdout
I'm trying to use github-release to download assets in a CI
environment, and it's more convenient to be able to pipe the assets
directly to tar/unzip than to have to write them to a file and call
tar separately. Adds a `--stdout` flag that will write downloaded
files to os.Stdout instead of to disk.