github-release: transparantly support pagination
Unfortunately, it doesn't fully do everything mentioned in
https://github.com/itchio/gothub/issues/3 yet. Namely, it doesn't "stop"
when something is found. So if you're often calling this thing on huge
repo's, you'll get rate-limited pretty soon. github-release(1) doesn't
handle rate-limiting in any other way than yielding an error.

A way to decrease the impact of rate-limiting is to supply per_page=100.
This commit doesn't do so because it would be massively unclean (I'd have to
parse the URI into a *url.URL, and append the per_page parameter, ugh *)
which this commit doesn't do bec.

Fixes #45

* The real solution is to use a http client wrapper that sends the AUTH
  token properly in the header, and refactor the way headers/query params
  are set.