upload: make uploads more robust
As described in issue #26, for some types of uploads the GitHub API tends to
be flaky and fail. The result is an inconsistent state (an asset in state
new, which doesn't seem useful for anything). Try hard to get rid of these
assets. This costs some more API calls, but I think people prefer
correctness over speed.

If that still doesn't cover all cases, the -R flag (for replacing assets)
should now work better (also for assets that weren't correctly uploaded).

This does not solve the issue where github-release apparently can't detect a
hung connection. Although it _should_ fix an upload failing on a retry of
this case.

Updates #26.